requirements.txt
pynetbox
python-dotenv